BC-4570 - move `FileContentElement` and its subcomponents to separate folder
