+++ src/components/feature-board/content-elements/ContentElementList.vue
@@ -31,11 +31,11 @@ import {
 	FileElementResponse,
 	RichTextElementResponse,
 } from "@/serverApi/v3";
-import { computed, defineComponent, PropType } from "vue";
+import { PropType, computed, defineComponent } from "vue";
 import { AnyContentElement } from "../types/ContentElement";
 import { ElementMove } from "../types/DragAndDrop";
-import FileContentElement from "./FileContentElement.vue";
 import RichTextContentElement from "./RichTextContentElement.vue";
+import FileContentElement from "./file-content-element/FileContentElement.vue";
 
 export default defineComponent({
 	name: "ContentElementList",
