+++ src/components/feature-board/content-elements/file-content-element/FileContentElement.vue
@@ -49,10 +49,10 @@ import {
 } from "@/fileStorageApi/v3";
 import { FileElementResponse } from "@/serverApi/v3";
 import { PropType, computed, defineComponent, onMounted } from "vue";
-import { useDeleteBoardNodeConfirmation } from "../shared/DeleteBoardNodeConfirmation.composable";
-import { useFileStorageApi } from "../shared/FileStorageApi.composable";
-import { useSelectedFile } from "../shared/SelectedFile.composable";
-import { useContentElementState } from "../state/ContentElementState.composable";
+import { useDeleteBoardNodeConfirmation } from "../../shared/DeleteBoardNodeConfirmation.composable";
+import { useFileStorageApi } from "../../shared/FileStorageApi.composable";
+import { useSelectedFile } from "../../shared/SelectedFile.composable";
+import { useContentElementState } from "../../state/ContentElementState.composable";
 import FileContentElementAlert from "./FileContentElementAlert.vue";
 import FileContentElementChips from "./FileContentElementChips.vue";
 import FileContentElementDisplay from "./FileContentElementDisplay.vue";
