+++ src/components/feature-board/content-elements/file-content-element/FileContentElementChips.vue
