+++ src/components/feature-board/content-elements/file-content-element/FileContentElementEdit.unit.ts
@@ -2,7 +2,7 @@ import { downloadFile } from "@/utils/fileHelper";
 import { I18N_KEY } from "@/utils/inject";
 import createComponentMocks from "@@/tests/test-utils/componentMocks";
 import { shallowMount } from "@vue/test-utils";
-import BoardMenuAction from "../shared/BoardMenuAction.vue";
+import BoardMenuAction from "../../shared/BoardMenuAction.vue";
 import FileContentElementEdit from "./FileContentElementEdit.vue";
 jest.mock("@/utils/fileHelper");
